db creacion e insercion inicial lista

+++ application/db/migrations/20210528045121_first_migrations.php
@@ -68,7 +68,7 @@ final class FirstMigrations extends AbstractMigration
             ->addColumn('trailer', 'string', [
                 'limit' => Constants::getLinkMax(),
                 'null' => true])
-            ->addColumn('valoracion', 'integer', ['null' => true])
+            ->addColumn('valoracion', 'float', ['null' => true])
             ->addColumn('activa', 'boolean', ['null' => false])
             ->addForeignKey('id_genero', 'generos', 'id')
             // ->addForeignKey('id_categoria', 'categorias', 'id')
